+++ packages/backend-common/src/database/connectors/sqlite3.ts
@@ -86,19 +86,6 @@ export function buildSqliteDatabaseConfig(
     overrides,
   );
 
-  // If we don't create an in-memory database, interpret the connection string
-  // as a directory that contains multiple sqlite files based on the database
-  // name.
-  const database = (config.connection as Knex.ConnectionConfig).database;
-  const sqliteConnection = config.connection as Knex.Sqlite3ConnectionConfig;
-
-  if (database && sqliteConnection.filename !== ':memory:') {
-    sqliteConnection.filename = path.join(
-      sqliteConnection.filename,
-      `${database}.sqlite`,
-    );
-  }
-
   return config;
 }