+++ packages/core-app-api/src/apis/implementations/auth/github/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-export * from './types';
+export type { GithubSession } from './types';
 export { default as GithubAuth } from './GithubAuth';