+++ plugins/catalog-backend/src/ingestion/processors/AzureDevOpsDiscoveryProcessor.test.ts
@@ -165,6 +165,7 @@ describe('AzureDevOpsDiscoveryProcessor', () => {
           type: 'url',
           target:
             'https://dev.azure.com/shopify/engineering/_git/backstage?path=/catalog-info.yaml',
+          presence: 'optional',
         },
         optional: true,
       });
@@ -174,6 +175,7 @@ describe('AzureDevOpsDiscoveryProcessor', () => {
           type: 'url',
           target:
             'https://dev.azure.com/shopify/engineering/_git/ios-app?path=/src/catalog-info.yaml',
+          presence: 'optional',
         },
         optional: true,
       });
@@ -211,6 +213,7 @@ describe('AzureDevOpsDiscoveryProcessor', () => {
           type: 'url',
           target:
             'https://dev.azure.com/shopify/engineering/_git/backstage?path=/catalog-info.yaml',
+          presence: 'optional',
         },
         optional: true,
       });
@@ -249,6 +252,7 @@ describe('AzureDevOpsDiscoveryProcessor', () => {
           type: 'url',
           target:
             'https://dev.azure.com/shopify/engineering/_git/backstage?path=/src/main/catalog.yaml',
+          presence: 'optional',
         },
         optional: true,
       });